requirements.txt
fastapi
uvicorn
openai
python-dotenv
webrtcvad
numpy
jinja2
python-multipart
streamlit